requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
openai-whisper==20231117
yt-dlp==2024.1.19
pydantic==2.5.3
